Switch from URI to Uri.
URI does not allow braces: {}, even though they are allowed in the spec.
It is also easier to build paths using a Uri instance.

Download/upload progress should be done with longs for large transfers.

+++ ion-sample/src/com/koushikdutta/ion/sample/ProgressBarUpload.java
@@ -69,7 +69,7 @@ public class ProgressBarUpload extends Activity {
                 // Updates to TextViews MUST happen on the UI thread.
                 .uploadProgressHandler(new ProgressCallback() {
                     @Override
-                    public void onProgress(int downloaded, int total) {
+                    public void onProgress(long downloaded, long total) {
                         uploadCount.setText("" + downloaded + " / " + total);
                     }
                 })
