+++ ion/src/com/koushikdutta/ion/Loader.java
@@ -21,7 +21,7 @@ public interface Loader {
         public static final int LOADED_FROM_CONDITIONAL_CACHE = 2;
         public static final int LOADED_FROM_NETWORK = 3;
 
-        public LoaderEmitter(DataEmitter emitter, int length, int loadedFrom, RawHeaders headers, AsyncHttpRequest request) {
+        public LoaderEmitter(DataEmitter emitter, long length, int loadedFrom, RawHeaders headers, AsyncHttpRequest request) {
             this.length = length;
             this.emitter = emitter;
             this.loadedFrom = loadedFrom;
@@ -29,11 +29,11 @@ public interface Loader {
             this.request = request;
         }
         DataEmitter emitter;
-        int length;
+        long length;
         public DataEmitter getDataEmitter() {
             return emitter;
         }
-        public int length() {
+        public long length() {
             return length;
         }
         int loadedFrom;
